% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{affine_transformation}
\alias{affine_transformation}
\title{Find Affine Transformation}
\usage{
affine_transformation(x, target)
}
\arguments{
\item{x}{A matrix of source coordinates.}

\item{target}{A matrix of target coordinates.}
}
\value{
A matrix of affine transformation.
}
\description{
This function calculates the affine transformation matrix to map a matrix of
source coordinates to a matrix of target coordinates. This function is originally
from the pscl package.
}
