% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{align_angles_to_direction}
\alias{align_angles_to_direction}
\title{Align Angles to a Given Direction}
\usage{
align_angles_to_direction(angles, direction)
}
\arguments{
\item{angles}{A vector of angles in radians.}

\item{direction}{A vector representing the direction to which angles should be aligned.}
}
\value{
A vector of angles aligned to the specified direction.
}
\description{
Align Angles to a Given Direction
}
