% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{align_to_principal_direction}
\alias{align_to_principal_direction}
\title{Align Angles to Principal Direction}
\usage{
align_to_principal_direction(angles)
}
\arguments{
\item{angles}{A vector of angles in radians.}
}
\value{
A vector of angles aligned to the principal direction.
}
\description{
This function aligns a set of angles to the principal direction derived from the angles.
Since we estimate the mean direction from the samples generated by Bingham distribution,
it is necessary to calculate the principal direction. This would mitigate the problem with
bimodal distribution of angles, which is by the nature of the Bingham distribution.
In case of single mode distribution, the principal direction is the sample average.
To ease the interpretation, the angles are aligned to the principal direction.
}
