% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{cartesian_to_polar}
\alias{cartesian_to_polar}
\title{Transform Cartesian Coordinates to Polar Coordinates}
\usage{
cartesian_to_polar(x, y)
}
\arguments{
\item{x}{The x coordinate (vector).}

\item{y}{The y coordinate (vector).}
}
\value{
A tibble of radius (r) and angle (theta) in polar coordinates.
}
\description{
This function converts Cartesian coordinates (x, y) to polar coordinates (radius, angle).
}
