% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{check_unanimous}
\alias{check_unanimous}
\title{Check Unanimous Votes}
\usage{
check_unanimous(votes, yea_code = 1, nay_code = 0)
}
\arguments{
\item{votes}{A matrix of votes (output of \code{pscl::rollcall(...)$votes}; recommended to use \code{filter_votes} first).}

\item{yea_code}{A vector of values in \code{votes} that are coded as yea (default: 1).}

\item{nay_code}{A vector of values in \code{votes} that are coded as nay (default: 0).}
}
\value{
A vector of roll call names where all legislators vote the same way.
}
\description{
This function checks if there are any roll calls where all legislators vote the same way.
}
