% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{Compute Cosine Similarity Between Two Vectors}
\usage{
cosine_similarity(a1, a2, b1, b2)
}
\arguments{
\item{a1}{The first scalar value of the first vector.}

\item{a2}{The second scalar value of the first vector.}

\item{b1}{The first scalar value of the second vector.}

\item{b2}{The second scalar value of the second vector.}
}
\value{
A scalar value representing the cosine similarity between the two vectors.
}
\description{
This function computes the cosine similarity between two vectors, each represented by two scalar values.
}
\examples{
cosine_similarity(1, 0, 0, 1) # Should return 0
cosine_similarity(1, 0, 1, 0) # Should return 1
}
