% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{filter_votes}
\alias{filter_votes}
\title{Preprocess Roll Call Data for Analysis}
\usage{
filter_votes(rollcall, lop = 0, minvotes = 0)
}
\arguments{
\item{rollcall}{A roll call object (output of \code{pscl::rollcall}).}

\item{lop}{The minimum level of support (or oppose) for a bill to be included in the analysis (default is 0).}

\item{minvotes}{The minimum number of votes a legislator must cast for them to be included in the analysis (default is 0).}
}
\value{
A list with two logical vectors: \code{legis} and \code{bills} indicating which legislators and bills should be included in the analysis.
}
\description{
This function preprocesses roll call data by filtering out votes based on specified criteria.
It drops votes according to the level of support or opposition for a bill and the minimum number of votes a legislator must cast.
}
