% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{find_constraints}
\alias{find_constraints}
\title{Find the Constraints in Legislative Voting Data}
\usage{
find_constraints(
  ideal,
  rollcall,
  pol_rc1,
  pol_rc2 = NULL,
  party_code_col = "party_code",
  left_party_code = 100,
  top_party_code = NULL,
  top_perc = 0.9,
  x2_threshold = 0,
  x2_threshold_second = NULL,
  vert_flip = FALSE,
  top_n = 0.1,
  as_list = FALSE
)
}
\arguments{
\item{ideal}{An ideal object (output of \code{pscl::ideal}).}

\item{rollcall}{A roll call object (output of \code{pscl::rollcall}; yea_code = 1, nay_code = 0).}

\item{pol_rc1}{A character representing the roll call ID for the most polarized bill in the horizontal dimension.}

\item{pol_rc2}{A character representing the roll call ID for the most polarized bill in the vertical dimension (optional).}

\item{party_code_col}{A character specifying the column name for party code.}

\item{left_party_code}{A party code for the party that should be on the left of the horizontal dimension.}

\item{top_party_code}{A party code for the party that should be on top of the vertical dimension (optional).}

\item{top_perc}{The percentage of legislators to be included based on their total number of votes cast.}

\item{x2_threshold}{The threshold for the vertical dimension for the first constraint (left of triangle).}

\item{x2_threshold_second}{The threshold for the vertical dimension for the second constraint (right of triangle).}

\item{vert_flip}{Whether to flip the second coordinate.}

\item{top_n}{A fraction of the data you want to select. For example, if top_n is set to 0.5, it means you want to select the top 50\% of the legislators.}

\item{as_list}{Boolean indicating whether to return the constraints as a list (default is FALSE).}
}
\value{
A tibble or list of legislator IDs along with their x1 and x2 coordinates based on the constraints.
}
\description{
This function identifies constraints in legislative voting data based on polarization,
party codes, and specified thresholds.
}
