% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{find_pol_rc_horizontal}
\alias{find_pol_rc_horizontal}
\title{Find Most Polarized Bills in Horizontal Dimension}
\usage{
find_pol_rc_horizontal(
  rollcall,
  party_code_col = "party_code",
  liberal_code = 100,
  conservative_code = 200,
  na_threshold = 0.3
)
}
\arguments{
\item{rollcall}{A roll call object (output of \code{pscl::rollcall}; yea_code = 1, nay_code = 0).}

\item{party_code_col}{A character string specifying the column name for party code (default: "party_code").}

\item{liberal_code}{Party code for the liberal party (default: 100).}

\item{conservative_code}{Party code for the conservative party (default: 200).}

\item{na_threshold}{The maximum ratio of missing data for a bill to be considered (default: 0.3).}
}
\value{
A character string representing the roll call ID of the most polarized bill.
}
\description{
This function identifies the most polarized bills based on voting patterns in a horizontal dimension from roll call data.
}
