% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{find_pol_rc_vertical}
\alias{find_pol_rc_vertical}
\title{Find Most Polarized Bills in Vertical Dimension}
\usage{
find_pol_rc_vertical(
  ideal,
  rollcall,
  pol_rc_horizontal = NULL,
  party_code_col = "party_code",
  liberal_code = 100,
  conservative_code = 200,
  na_threshold = 0.3,
  lop_threshold = 0.1
)
}
\arguments{
\item{ideal}{An ideal object (output of \code{pscl::ideal}).}

\item{rollcall}{A roll call object (output of \code{pscl::rollcall}; yea_code = 1, nay_code = 0).}

\item{pol_rc_horizontal}{A character vector of roll call IDs for the most polarized bill in the horizontal dimension (default: NULL, will be computed if not provided).}

\item{party_code_col}{A character string specifying the column name for party code (default: "party_code").}

\item{liberal_code}{Party code for the liberal party (default: 100).}

\item{conservative_code}{Party code for the conservative party (default: 200).}

\item{na_threshold}{The maximum ratio of missing data for a bill to be considered (default: 0.3).}

\item{lop_threshold}{The minimum level of support (or oppose) for a bill to be considered (default: 0.1).}
}
\value{
A character string representing the roll call ID of the most polarized bill in the vertical dimension.
}
\description{
This function identifies the most polarized bills in the vertical dimension based on ideal point estimates and roll call data.
}
