% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\name{generate_dynamic_data}
\alias{generate_dynamic_data}
\title{Generate Synthetic Roll Call Data (Dynamic)}
\usage{
generate_dynamic_data(
  seed = 2138,
  n = 30,
  m = 270,
  n_terms = 3,
  k_ls = c(3, 3, 3),
  kappa = 5,
  rho = 10,
  a = 0.01,
  b = 0.001,
  theta_ls = NULL,
  X = NULL,
  lop = 0,
  minvotes = 20,
  t_ls = replicate(30, 1:3, simplify = FALSE),
  missingness_rate = 0.001,
  fixed_sign = FALSE
)
}
\arguments{
\item{seed}{Random seed for reproducibility.}

\item{n}{Number of legislators.}

\item{m}{Number of bills.}

\item{n_terms}{Number of terms.}

\item{k_ls}{Number of topics in each term.}

\item{kappa}{Standard deviation of w.}

\item{rho}{Concentration parameter of von Mises distribution.}

\item{a}{Hyperparameter for the prior of rho (0 < b < a).}

\item{b}{Hyperparameter for the prior of theta (0 < b < a).}

\item{theta_ls}{Issue vectors (in angle), optional.}

\item{X}{Legislator ideal points (in 2D), optional.}

\item{lop}{Lower bound of proportion of votes, used in filtering.}

\item{minvotes}{Minimum number of votes required for a legislator to be included.}

\item{t_ls}{A list of terms for each legislator.}

\item{missingness_rate}{Rate of missingness in the data.}

\item{fixed_sign}{Logical, if TRUE, the yea_to_nay direction is fixed.}
}
\value{
A list containing 'data' for general use and 'stan' for inputs to a Stan model.
}
\description{
This function generates synthetic roll call data for a specified number of legislators and bills,
along with the necessary inputs for a Stan model. It simulates issue vectors, roll call vectors,
difficulty parameters, and legislator ideal points.
}
