% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_congress_years}
\alias{get_congress_years}
\title{Get Congress Years}
\usage{
get_congress_years(congress_number)
}
\arguments{
\item{congress_number}{The number of the U.S. Congress.}
}
\value{
A string representing the years during which the specified Congress was in session.
}
\description{
This function calculates the years during which a specified number of the U.S. Congress was in session.
The calculation is based on the fact that the 1st Congress started in 1789, and each Congress spans two years.
}
\examples{
get_congress_years(117) # Returns "2021 - 2022" for the 117th Congress
}
