% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic.R
\name{get_dynamic_constraints}
\alias{get_dynamic_constraints}
\title{Generate dynamic version of constraints (for visualization of multidimensional ideal points)}
\usage{
get_dynamic_constraints(dynamic_stan_fit, dynamic_stan_input, anchors_name)
}
\arguments{
\item{dynamic_stan_fit}{A stan fit object.}

\item{dynamic_stan_input}{A dynamic_stan_input object.}

\item{anchors_name}{A character vector specifying the name of each anchor.}
}
\value{
A list of constraints for visualization.
}
\description{
Generate dynamic version of constraints (for visualization of multidimensional ideal points)
}
