% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{get_principal_direction}
\alias{get_principal_direction}
\title{Get Principal Direction in Polar Coordinates}
\usage{
get_principal_direction(angles, return_angle = FALSE)
}
\arguments{
\item{angles}{A vector of angles in radians.}

\item{return_angle}{A logical value indicating whether to return the angle of the principal direction}
}
\value{
A vector representing the principal direction in Cartesian coordinates, or the angle of the principal direction if \code{return_angle} is TRUE.
}
\description{
Get Principal Direction in Polar Coordinates
}
