% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issueipe-package.R
\docType{package}
\name{issueipe-package}
\alias{issueipe-package}
\alias{issueipe}
\title{The 'issueipe' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.3. https://mc-stan.org
}
