% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issueirt-package.R
\docType{package}
\name{issueirt-package}
\alias{issueirt-package}
\title{The 'issueirt' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.3. https://mc-stan.org
}
\author{
\strong{Maintainer}: Sooahn Shin \email{sooahnshin@fas.harvard.edu} (\href{https://orcid.org/0000-0001-6213-2197}{ORCID})

}
