% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issueirt_stan.R
\name{issueirt_stan}
\alias{issueirt_stan}
\title{Run Stan Model}
\usage{
issueirt_stan(
  data,
  init,
  chains = 2,
  iter = 10,
  warmup = 5,
  thin = 1,
  seed = 1,
  ...
)
}
\arguments{
\item{data}{A list of data for Stan.}

\item{init}{A list of initial values for Stan.}

\item{chains}{Number of chains (default: 2).}

\item{iter}{Number of iterations (default: 10).}

\item{warmup}{Number of warmup iterations (default: 5).}

\item{thin}{Thinning rate (default: 1).}

\item{seed}{Random seed (default: 1).}

\item{...}{Other arguments to be passed to \code{rstan::sampling}.}
}
\value{
A Stan fit object.
}
\description{
This function runs a Stan model using the \code{rstan::sampling} function to
estimate issue specific ideal points.
It allows for specifying various parameters for the Stan model execution.
}
