% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic.R
\name{make_dynamic_rollcall}
\alias{make_dynamic_rollcall}
\title{Generate dynamic version of rollcall object}
\usage{
make_dynamic_rollcall(
  votes_list,
  issue_list,
  legis_list,
  colname_legis,
  colname_party = NULL,
  bills_list = NULL,
  colname_bills = NULL,
  term_name = NULL,
  left_code = 1,
  right_code = 0,
  missing_code = NA,
  notInLegis_code = 9
)
}
\arguments{
\item{votes_list}{A list of roll call votes data. Each element of the list is a data frame where rows are legislators and columns are roll call votes.}

\item{issue_list}{A list of issue data. Each element of the list is a data frame where each row is an issue and columns are issue information.}

\item{legis_list}{A list of legislator data. Each element of the list is a data frame where each row is a legislator and columns are legislator information.}

\item{colname_legis}{A character string specifying the column name of legislator ID in legis_list.}

\item{colname_party}{A character string specifying the column name of party affiliation in legis_list. If NULL, party affiliation is not included in the rollcall object.}

\item{bills_list}{A list of bill data. Each element of the list is a data frame where each row is a bill and columns are bill information.}

\item{colname_bills}{A character string specifying the column name of bill ID in bills_list. If NULL, bill ID is not included in the rollcall object.}

\item{term_name}{A character vector specifying the name of each term. If NULL, the default term name is "T1", "T2", ..., "Tn".}

\item{left_code}{A numeric value specifying the code for "left-wing" votes.}

\item{right_code}{A numeric value specifying the code for "right-wing" votes.}

\item{missing_code}{A numeric value specifying the code for missing votes.}

\item{notInLegis_code}{A numeric value specifying the code for legislators who did not vote.}
}
\value{
A dynamic_rollcall object.
}
\description{
Generate dynamic version of rollcall object
}
