% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic.R
\name{make_dynamic_stan_input}
\alias{make_dynamic_stan_input}
\title{Generate dynamic version of stan input object}
\usage{
make_dynamic_stan_input(
  dynamic_rollcall,
  ideal,
  a = 0.01,
  b = 0.001,
  rho_init = 10,
  sd_dynamic = 0.01
)
}
\arguments{
\item{dynamic_rollcall}{A dynamic_rollcall object.}

\item{ideal}{An ideal object.}

\item{a}{hyperparameter for the prior of rho 0<b<a}

\item{b}{hyperparameter for the prior of theta 0<b<a}

\item{rho_init}{initial value for rho}

\item{sd_dynamic}{standard deviation for dynamic prior}
}
\value{
A dynamic_stan_input object.
}
\description{
Generate dynamic version of stan input object
}
