% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{make_issue_code}
\alias{make_issue_code}
\title{Recode Issue Codes for Stan}
\usage{
make_issue_code(issue_code_vec, levels = NULL)
}
\arguments{
\item{issue_code_vec}{A vector of issue codes, expected to be a character vector.}

\item{levels}{An optional vector specifying the order of issue labels. If provided, it must be a character vector
and have a length equal to the number of unique values in \code{issue_code_vec}.}
}
\value{
A list containing the recoded issue code vector and a codebook (as a tibble) mapping the codes to labels.
}
\description{
This function recodes a vector of issue codes, typically as part of pre-processing for Stan model input.
It also generates a codebook for the recoded issue codes.
}
