% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{make_missing_indicator}
\alias{make_missing_indicator}
\title{Generate Label for Missing Indicator from Roll Call Data}
\usage{
make_missing_indicator(rollcall)
}
\arguments{
\item{rollcall}{A roll call object (output of \code{pscl::rollcall}).}
}
\value{
A vector of labels for missing indicators, formatted as 'legis_id bill_id'.
}
\description{
This function creates labels for missing indicators in roll call data.
It identifies missing votes and assigns labels based on legislator and bill IDs.
}
