% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{make_posterior_summary}
\alias{make_posterior_summary}
\title{Generate Posterior Summary of Stan Output}
\usage{
make_posterior_summary(
  stan_fit,
  issue_label = NULL,
  rc_label = NULL,
  legis_label = NULL,
  missing_label = NULL
)
}
\arguments{
\item{stan_fit}{A Stan fit object.}

\item{issue_label}{A vector of issue labels.}

\item{rc_label}{A vector of roll call labels.}

\item{legis_label}{A vector of legislator labels.}

\item{missing_label}{A vector of missing indicator labels.}
}
\value{
A list of tibbles containing the posterior summary for various model parameters, including theta,
theta's auxiliary variables, u, u's auxiliary variables, x, other bill parameters, and imputed missing data.
}
\description{
This function processes the output from a Stan model and generates a summary of posterior distributions.
It organizes the summary into different categories based on the parameters of the model.
}
