% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{make_posterior_summary_postprocessed}
\alias{make_posterior_summary_postprocessed}
\title{Generate Posterior Summary of Post-Processed Stan Output}
\usage{
make_posterior_summary_postprocessed(
  stan_fit,
  constraints,
  issue_label = NULL,
  rc_label = NULL,
  legis_label = NULL,
  missing_label = NULL
)
}
\arguments{
\item{stan_fit}{A Stan fit object.}

\item{constraints}{A list of constraints for post-processing.}

\item{issue_label}{A vector of issue labels (optional).}

\item{rc_label}{A vector of roll call labels (optional).}

\item{legis_label}{A vector of legislator labels (optional).}

\item{missing_label}{A vector of missing indicator labels (optional).}
}
\value{
A list of tibbles containing posterior summaries:
\itemize{
\item theta: Posterior summary of theta.
\item theta_aux: Posterior summary of auxiliary variables of theta (x_coord_theta, y_coord_theta, r_theta).
\item u: Posterior summary of u.
\item u_aux: Posterior summary of auxiliary variables of u (x_coord_u, y_coord_u, r_u).
\item x: Posterior summary of x.
\item other_bill: Posterior summary of other bill parameters (alpha, w, rho).
\item y_imputed: Posterior summary of imputed missing data.
\item theta_postprocessed: Posterior summary of theta (post-processed).
\item u_postprocessed: Posterior summary of u (post-processed).
\item x_postprocessed: Posterior summary of x (post-processed).
}
}
\description{
This function generates posterior summaries of post-processed Stan output, including samples of x, u, and theta. It applies affine transformations based on constraints and provides posterior summaries for various parameters.
}
