% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{make_stan_input}
\alias{make_stan_input}
\title{Generate data and initial values for stan}
\usage{
make_stan_input(
  issue_code_vec,
  rollcall,
  ideal,
  a = 0.01,
  b = 0.001,
  rho_init = 10,
  fixed_sign = FALSE
)
}
\arguments{
\item{issue_code_vec}{A vector of issue codes}

\item{rollcall}{A roll call object (output of \code{pscl::rollcall})}

\item{ideal}{An ideal object (output of \code{pscl::ideal})}

\item{a}{hyperparameter for the prior of rho 0<b<a}

\item{b}{hyperparameter for the prior of theta 0<b<a}

\item{rho_init}{initial value for rho}

\item{fixed_sign}{Logical, if TRUE, the yea_to_nay direction is fixed.}
}
\value{
A list of data and initial values for Stan
}
\description{
This function generates data and initial values for Stan modeling.
}
