% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issueirt.R
\name{plot.issueirt_fit}
\alias{plot.issueirt_fit}
\title{Plot Method for issueirt_fit Objects}
\usage{
\method{plot}{issueirt_fit}(x, type = c("ideal_points", "axes", "issue_specific"), ...)
}
\arguments{
\item{x}{An \code{issueirt_fit} object.}

\item{type}{Type of plot to create. Options are "ideal_points" (default),
"axes", or "issue_specific".}

\item{...}{Additional arguments passed to the underlying plot functions.}
}
\value{
A ggplot object.
}
\description{
Plot Method for issueirt_fit Objects
}
