% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_ideal}
\alias{plot_ideal}
\title{Plot the Ideal Point Estimates}
\usage{
plot_ideal(
  ideal_point_1d,
  ideal_point_2d,
  group,
  p.title = NULL,
  breaks.group = NULL,
  values.shape = NULL,
  values.color = NULL
)
}
\arguments{
\item{ideal_point_1d}{A vector of ideal point estimates in the horizontal dimension.}

\item{ideal_point_2d}{A vector of ideal point estimates in the vertical dimension.}

\item{group}{A vector of group membership.}

\item{p.title}{The title of the plot (optional).}

\item{breaks.group}{A vector of group names for legend breaks (optional).}

\item{values.shape}{A vector of shape values for each group (optional).}

\item{values.color}{A vector of color values for each group (optional).}
}
\value{
A ggplot object representing the ideal point estimates.
}
\description{
This function creates a ggplot object to visualize ideal point estimates in a two-dimensional space, distinguishing groups by color and shape.
}
