% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_issueaxis}
\alias{plot_issueaxis}
\title{Plot the Issue Specific Axes}
\usage{
plot_issueaxis(
  stan_input,
  posterior_summary,
  p.title = NULL,
  group = NULL,
  breaks.group = NULL,
  values.shape = NULL,
  values.color = NULL
)
}
\arguments{
\item{stan_input}{An issueirt_stan_input object.}

\item{posterior_summary}{An issueirt_posterior_summary object.}

\item{p.title}{The title of the plot (optional).}

\item{group}{A vector of group membership (optional).}

\item{breaks.group}{A vector of group names for legend breaks (optional).}

\item{values.shape}{A vector of shape values for each group (optional).}

\item{values.color}{A vector of color values for each group (optional).}
}
\value{
A list of ggplot objects representing the issue specific axes.
}
\description{
This function creates a ggplot object to visualize issue specific axes using the stan input and the output of the \code{make_posterior_summary_postprocessed} function.
}
