% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_issueirt}
\alias{plot_issueirt}
\title{Plot the Issue Specific Ideal Points}
\usage{
plot_issueirt(
  issueirt,
  p.title = NULL,
  breaks.group = NULL,
  values.shape = NULL,
  values.color = NULL,
  alpha = 1,
  size = 3
)
}
\arguments{
\item{issueirt}{An issueirt_ideal_points object.}

\item{p.title}{The title of the plot (optional).}

\item{breaks.group}{A vector of group names for legend breaks (optional).}

\item{values.shape}{A vector of shape values for each group (optional).}

\item{values.color}{A vector of color values for each group (optional).}

\item{alpha}{The transparency of the points (optional).}

\item{size}{The size of the points (optional).}
}
\value{
A ggplot object representing the issue specific ideal points.
}
\description{
This function creates a ggplot object to visualize issue specific ideal points using the output of the \code{get_ideal_points} function.
}
