% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{polar_to_cartesian}
\alias{polar_to_cartesian}
\title{Transform Polar Coordinates to Cartesian Coordinates}
\usage{
polar_to_cartesian(r, theta)
}
\arguments{
\item{r}{The radius (vector).}

\item{theta}{The angle (vector).}
}
\value{
A tibble of x and y coordinates in Cartesian coordinates.
}
\description{
This function converts polar coordinates (radius, angle) to Cartesian coordinates (x, y).
}
