% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{post_process}
\alias{post_process}
\title{Post-process Function Based on pscl Package}
\usage{
post_process(stan_fit, constraints, legis_label, as_mcmc = FALSE)
}
\arguments{
\item{stan_fit}{A stan fit object.}

\item{constraints}{A list of constraints.}

\item{legis_label}{A vector of legislator labels.}

\item{as_mcmc}{Whether to return the posterior samples as mcmc objects.}
}
\value{
A list of posterior samples of x, u, and theta that are post-processed.
}
\description{
This function post-processes the posterior samples from a Stan fit object to obtain
samples of x, u, and theta. It also applies affine transformations based on constraints.
}
