% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{postprocess_xbar}
\alias{postprocess_xbar}
\title{Rotate and Normalize Posterior Mean of Bayesian IRT}
\usage{
postprocess_xbar(
  ideal,
  rollcall,
  pol_rc1,
  pol_rc2 = NULL,
  party_code_col = "party_code",
  left_party_code = 100,
  top_party_code = NULL
)
}
\arguments{
\item{ideal}{An ideal object (output of \code{pscl::ideal}).}

\item{rollcall}{A roll call object (output of \code{pscl::rollcall}; yea_code = 1, nay_code = 0).}

\item{pol_rc1}{A character representing the roll call ID for the most polarized bill in the horizontal dimension.}

\item{pol_rc2}{A character representing the roll call ID for the most polarized bill in the vertical dimension (optional).}

\item{party_code_col}{A character specifying the column name for party code.}

\item{left_party_code}{A party code for the party that should be on the left of the horizontal dimension.}

\item{top_party_code}{A party code for the party that should be on top of the vertical dimension (optional).}
}
\value{
A matrix of the posterior mean of Bayesian IRT with two columns (x1 and x2), each normalized to have mean 0 and standard deviation 1.
}
\description{
This function rotates and normalizes the posterior mean of Bayesian Item Response Theory (IRT)
based on specified criteria. It adjusts the orientation of the IRT space based on the most
polarized bills and ensures that specified parties are correctly positioned in the rotated space.
}
