% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issueirt.R
\name{print.issueirt_fit}
\alias{print.issueirt_fit}
\title{Print Method for issueirt_fit Objects}
\usage{
\method{print}{issueirt_fit}(x, ...)
}
\arguments{
\item{x}{An \code{issueirt_fit} object.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for issueirt_fit Objects
}
