% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{process_data}
\alias{process_data}
\title{Process Data (for Internal Use of \code{make_posterior_summary_postprocessed})}
\usage{
process_data(data, parameter, index_label, label_data = NULL, dimension = NULL)
}
\arguments{
\item{data}{A matrix of data.}

\item{parameter}{A character of parameter name.}

\item{index_label}{A character of index label.}

\item{label_data}{A vector of labels.}

\item{dimension}{A vector of dimension.}
}
\value{
A tibble of posterior summary.
}
\description{
This function processes data for generating posterior summaries.
}
