% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{recode_votes}
\alias{recode_votes}
\title{Recode Roll Call Votes (Optional)}
\usage{
recode_votes(
  votes,
  party_code,
  liberal_code = 100,
  conservative_code = 200,
  yea = 1,
  nay = 0,
  missing = c(NA, 2),
  na_threshold = 0.5
)
}
\arguments{
\item{votes}{A matrix of votes (output of \code{pscl::rollcall(...)$votes}; recommended to use \code{filter_votes} first).}

\item{party_code}{A vector of party codes for each legislator.}

\item{liberal_code}{Party code for the liberal party (default: 100).}

\item{conservative_code}{Party code for the conservative party (default: 200).}

\item{yea}{A vector of values in \code{votes} that are coded as yea (default: 1).}

\item{nay}{A vector of values in \code{votes} that are coded as nay (default: 0).}

\item{missing}{A vector of values in \code{votes} that are coded as missing (default: c(NA, 2)).}

\item{na_threshold}{The maximum ratio of missing data for a bill to be recoded (default: 0.5).}
}
\value{
A matrix of recoded votes. 1 indicates a liberal majority or conservative minority, and 0 indicates a conservative majority or liberal minority.
}
\description{
This function preprocesses roll call vote data by recoding votes to prevent sign flipping.
It is designed to work with matrices of votes, those obtained from \code{pscl::rollcall} for example.
You may use this only when you fit the model using von Mises-Fisher distribution
instead of Bingham distribution (default).
}
