% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{rotate_matrix}
\alias{rotate_matrix}
\title{Rotate Matrix by Angle Specified with Slope and Origin}
\usage{
rotate_matrix(M, slope = 1, origin = c(0, 0))
}
\arguments{
\item{M}{A matrix of x and y coordinates of the vectors to be rotated (each row is a vector of length 2).}

\item{slope}{The slope for the rotation (default is 1).}

\item{origin}{A vector of x and y coordinates of the origin for the rotation (default is c(0, 0)).}
}
\value{
A matrix of x and y coordinates of the rotated vectors.
}
\description{
This function rotates a matrix of vectors by an angle determined by a given slope and origin. Each row of the matrix represents a vector.
}
\examples{
M <- matrix(c(1, 0, 0, 1), ncol = 2) # Two vectors: (1,0) and (0,1)
rotate_matrix(M, slope = 1) # Rotates both vectors
}
