% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{rotate_vector}
\alias{rotate_vector}
\title{Rotate Vector by Angle Specified with Slope and Origin}
\usage{
rotate_vector(x, y, slope = 1, origin = c(0, 0))
}
\arguments{
\item{x}{The x coordinate of the vector.}

\item{y}{The y coordinate of the vector.}

\item{slope}{The slope for the rotation (default is 1).}

\item{origin}{A vector of x and y coordinates of the origin for the rotation (default is c(0, 0)).}
}
\value{
A matrix of x and y coordinates of the rotated vector.
}
\description{
This function rotates a vector by an angle determined by a given slope and origin.
}
\examples{
# Rotates the vector (1,0) by the angle corresponding to slope = 1
# around the origin
rotate_vector(1, 0, slope = 1, origin = c(0, 0))
}
