% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issueirt_dynamic.R
\name{summary.issueirt_dynamic_fit}
\alias{summary.issueirt_dynamic_fit}
\title{Summary Method for issueirt_dynamic_fit Objects}
\usage{
\method{summary}{issueirt_dynamic_fit}(object, ...)
}
\arguments{
\item{object}{An \code{issueirt_dynamic_fit} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A list containing summary information.
}
\description{
Summary Method for issueirt_dynamic_fit Objects
}
