% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issueirt.R
\name{summary.issueirt_fit}
\alias{summary.issueirt_fit}
\title{Summary Method for issueirt_fit Objects}
\usage{
\method{summary}{issueirt_fit}(object, ...)
}
\arguments{
\item{object}{An \code{issueirt_fit} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A list containing summary information.
}
\description{
Summary Method for issueirt_fit Objects
}
