% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{update_recode_votes}
\alias{update_recode_votes}
\title{Fix sign-flip based on the fit (Optional)}
\usage{
update_recode_votes(votes, stan_input, posterior_summary)
}
\arguments{
\item{votes}{A matrix of votes (output of \code{recode_votes}). It may only contain 0, 1, and NA.}

\item{stan_input}{An issueirt_stan_input or dynamic_stan_input object.}

\item{posterior_summary}{An issueirt_posterior_summary object.}
}
\value{
A matrix of updated votes. In case of dynamic model, it returns a list of matrices of updated votes.
}
\description{
Update the votes based on the cosine similarity between the estimated direction
of rollcall and the estimated direction of issue-specific axes. If the cosine similarity is negative,
it is a sign of sign-flip. This function updates the votes to fix the sign-flip.
You may use this only when you fit the model using von Mises-Fisher distribution
instead of Bingham distribution (default).
}
