% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us1890s_bills}
\alias{us1890s_bills}
\title{Roll-Calls in U.S. House of Representatives, 1890s (52nd to 54th)}
\format{
A list of three \code{data.frame}s, one for each Congress (52nd to 54th).
}
\usage{
data(us1890s_bills)
}
\description{
This dataset contains information on roll-call votes in the U.S. House of Representatives from the 52nd to the 54th Congress.
}
\references{
Lewis, Jeffrey B., Keith Poole, Howard Rosenthal, Adam Boche, Aaron Rudkin, and Luke Sonnet. (2024)
\emph{Voteview: Congressional Roll-Call Votes Database.}
\url{https://voteview.com/}
}
\seealso{
\code{\link{us1890s_votes}}, \code{\link{us1890s_legis}}, \code{\link{us1890s_issue}}
}
\keyword{dataset}
