% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us1890s_issue}
\alias{us1890s_issue}
\title{Issue Codes in U.S. House of Representatives, 1890s (52nd to 54th)}
\format{
A list of three \code{data.frame}s, one for each Congress (52nd to 54th).
}
\usage{
data(us1890s_issue)
}
\description{
This dataset contains information on issue codes in the U.S. House of Representatives from the 52nd to the 54th Congress.
}
\references{
Bateman, David A, Ira Katznelson and John S Lapinski. (2022)
"Issue codings for roll call in the U.S. House of Representatives and Senate between 1877 and 2011"
\emph{American Institutions Project.}
\url{http://www.davidalexbateman.net/congressional-data.html.}
}
\seealso{
\code{\link{us1890s_votes}}, \code{\link{us1890s_legis}}, \code{\link{us1890s_bills}}
}
\keyword{dataset}
