% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us1890s_votes}
\alias{us1890s_votes}
\title{Roll Call Votes in U.S. House of Representatives, 1890s (52nd to 54th)}
\format{
A list of three \code{data.frame}s, one for each Congress (52nd to 54th).
Each data frame has rows representing representatives (ICPSR) and columns representing votes (roll number).
}
\usage{
data(us1890s_votes)
}
\description{
This dataset contains roll call votes in the U.S. House of Representatives from the 52nd to the 54th Congress.
Each element in the list is a data frame where rows correspond to representatives (identified by ICPSR) and columns correspond to votes (identified by roll number).
}
\references{
Lewis, Jeffrey B., Keith Poole, Howard Rosenthal, Adam Boche, Aaron Rudkin, and Luke Sonnet. (2024)
\emph{Voteview: Congressional Roll-Call Votes Database.}
\url{https://voteview.com/}
}
\seealso{
\code{\link{us1890s_legis}}, \code{\link{us1890s_bills}}, \code{\link{us1890s_issue}}
}
\keyword{dataset}
