% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\name{generate_data}
\alias{generate_data}
\title{Generate Synthetic Roll Call Data}
\usage{
generate_data(
  seed = 2138,
  n = 20,
  m = 100,
  k = 5,
  kappa = 5,
  rho = 10,
  a = 0.01,
  b = 0.001,
  theta = NULL,
  X = NULL,
  lop = 0,
  minvotes = 20
)
}
\arguments{
\item{seed}{Random seed for reproducibility.}

\item{n}{Number of legislators.}

\item{m}{Number of bills.}

\item{k}{Number of topics.}

\item{kappa}{Standard deviation of w.}

\item{rho}{Concentration parameter of von Mises distribution.}

\item{a}{Hyperparameter for the prior of rho (0 < b < a).}

\item{b}{Hyperparameter for the prior of theta (0 < b < a).}

\item{theta}{Issue vectors (in angle), optional.}

\item{X}{Legislator ideal points (in 2D), optional.}

\item{lop}{Lower bound of proportion of votes, used in filtering.}

\item{minvotes}{Minimum number of votes required for a legislator to be included.}
}
\value{
A list containing 'data' for general use and 'stan' for inputs to a Stan model.
}
\description{
This function generates synthetic roll call data for a specified number of legislators and bills,
along with the necessary inputs for a Stan model. It simulates issue vectors, roll call vectors,
difficulty parameters, and legislator ideal points.
}
