% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{get_ideal_points}
\alias{get_ideal_points}
\title{Generate Issue Specific Ideal Points}
\usage{
get_ideal_points(
  stan_fit,
  issue_label = NULL,
  legis_label = NULL,
  legis_group = NULL,
  save.samples = FALSE
)
}
\arguments{
\item{stan_fit}{A Stan fit object.}

\item{issue_label}{A vector of issue labels.}

\item{legis_label}{A vector of legislator labels.}

\item{legis_group}{A vector of legislator group labels.}

\item{save.samples}{A logical value indicating whether to save samples (optional).}
}
\value{
A tibble of posterior summary for issue-specific ideal points.
}
\description{
This function generates issue-specific ideal points based on the Stan fit object.
}
