% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issueirt-package.R
\docType{_PACKAGE}
\name{issueirt-package}
\alias{issueirt-package}
\alias{issueirt}
\title{The 'issueirt' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.3. https://mc-stan.org
}
